Add api to accept data requests
In this PR I change how the accepted_at is used. It was in a weird place
because right now the owner could create a request that has it set
automatically. So basically the code was not really used. I now use it
to let the owner approve the requests made through IRIS. This also
made it possible to remove the dead code :)

references [#125](https://github.com/railslove/recover-backlog/issues/125)